Internal trace module refactor
Simplify module structure by combining small single-type modules into
their parent module.
