--- opentelemetry/src/trace/id_generator.rs
@@ -1,12 +1,0 @@
-//! # OpenTelemetry Id Generator Interface
-use crate::trace::{SpanId, TraceId};
-use std::fmt;
-
-/// Interface for generating IDs
-pub trait IdGenerator: Send + Sync + fmt::Debug {
-    /// Generate a new `TraceId`
-    fn new_trace_id(&self) -> TraceId;
-
-    /// Generate a new `SpanId`
-    fn new_span_id(&self) -> SpanId;
-}
